Previous onboarding DaxDialogs removed
