+++ app/src/main/java/com/duckduckgo/app/browser/commands/Command.kt
@@ -34,8 +34,7 @@ import com.duckduckgo.app.browser.history.NavigationHistoryEntry
 import com.duckduckgo.app.browser.model.BasicAuthenticationCredentials
 import com.duckduckgo.app.browser.model.BasicAuthenticationRequest
 import com.duckduckgo.app.browser.viewstate.SavedSiteChangedViewState
-import com.duckduckgo.app.cta.ui.Cta
-import com.duckduckgo.app.cta.ui.ExperimentOnboardingDaxDialogCta
+import com.duckduckgo.app.cta.ui.OnboardingDaxDialogCta
 import com.duckduckgo.app.fire.fireproofwebsite.data.FireproofWebsiteEntity
 import com.duckduckgo.app.survey.model.Survey
 import com.duckduckgo.autofill.api.domain.app.LoginCredentials
@@ -200,11 +199,6 @@ sealed class Command {
 
     class ShowEmailProtectionChooseEmailPrompt(val address: String) : Command()
     object ShowEmailProtectionInContextSignUpPrompt : Command()
-    sealed class DaxCommand : Command() {
-        object FinishPartialTrackerAnimation : DaxCommand()
-        class HideDaxDialog(val cta: Cta) : DaxCommand()
-    }
-
     class CancelIncomingAutofillRequest(val url: String) : Command()
     object LaunchAutofillSettings : Command()
     class EditWithSelectedQuery(val query: String) : Command()
@@ -231,12 +225,11 @@ sealed class Command {
     data class ScreenLock(val data: JsCallbackData) : Command()
     object ScreenUnlock : Command()
     data object ShowFaviconsPrompt : Command()
-    data class SetBrowserBackground(val backgroundRes: Int) : Command()
     data class ShowSSLError(val handler: SslErrorHandler, val error: SslErrorResponse) : Command()
     data object HideSSLError : Command()
     class LaunchScreen(
         val screen: String,
         val payload: String,
     ) : Command()
-    data class HideExperimentOnboardingDialog(val experimentCta: ExperimentOnboardingDaxDialogCta) : Command()
+    data class HideExperimentOnboardingDialog(val experimentCta: OnboardingDaxDialogCta) : Command()
 }
