+++ app/src/main/java/com/duckduckgo/app/browser/omnibar/animations/BrowserLottieTrackersAnimatorHelper.kt
@@ -57,9 +57,6 @@ class BrowserLottieTrackersAnimatorHelper @Inject constructor(
     private lateinit var cookieViewBackground: View
     private var cookieCosmeticHide: Boolean = false
 
-    private var runPartialAnimation: Boolean = false
-    private var completePartialAnimation: Boolean = false
-
     private var enqueueCookiesAnimation = false
     private var isCookiesAnimationRunning = false
     private var hasCookiesAnimationBeenCanceled = false
@@ -69,16 +66,14 @@ class BrowserLottieTrackersAnimatorHelper @Inject constructor(
 
     override fun startTrackersAnimation(
         context: Context,
-        shouldRunPartialAnimation: Boolean,
         shieldAnimationView: LottieAnimationView,
         trackersAnimationView: LottieAnimationView,
         omnibarViews: List<View>,
         entities: List<Entity>?,
     ) {
         if (isCookiesAnimationRunning) return // If cookies animation is running let it finish to avoid weird glitches with the other animations
-        if (trackersAnimationView.isAnimating || this.runPartialAnimation) return
+        if (trackersAnimationView.isAnimating) return
 
-        this.runPartialAnimation = shouldRunPartialAnimation
         this.trackersAnimation = trackersAnimationView
         this.shieldAnimation = shieldAnimationView
 
@@ -103,17 +98,13 @@ class BrowserLottieTrackersAnimatorHelper @Inject constructor(
             this.addAnimatorListener(
                 object : AnimatorListener {
                     override fun onAnimationStart(animation: Animator) {
-                        if (completePartialAnimation) return
                         animateOmnibarOut(omnibarViews).start()
                     }
 
                     override fun onAnimationEnd(animation: Animator) {
-                        if (!runPartialAnimation) {
-                            animateOmnibarIn(omnibarViews).start()
-                            completePartialAnimation = false
-                            tryToStartCookiesAnimation(context, omnibarViews)
-                            listener?.onAnimationFinished()
-                        }
+                        animateOmnibarIn(omnibarViews).start()
+                        tryToStartCookiesAnimation(context, omnibarViews)
+                        listener?.onAnimationFinished()
                     }
 
                     override fun onAnimationCancel(animation: Animator) {
@@ -124,13 +115,8 @@ class BrowserLottieTrackersAnimatorHelper @Inject constructor(
                 },
             )
 
-            if (runPartialAnimation) {
-                this.setMaxProgress(0.5f)
-                shieldAnimationView.setMaxProgress(0.5f)
-            } else {
-                this.setMaxProgress(1f)
-                shieldAnimationView.setMaxProgress(1f)
-            }
+            this.setMaxProgress(1f)
+            shieldAnimationView.setMaxProgress(1f)
             shieldAnimationView.playAnimation()
             this.playAnimation()
         }
@@ -149,7 +135,7 @@ class BrowserLottieTrackersAnimatorHelper @Inject constructor(
         this.cookieView = cookieAnimationView
         this.cookieCosmeticHide = cookieCosmeticHide
 
-        if (this.trackersAnimation?.isAnimating != true && !runPartialAnimation) {
+        if (this.trackersAnimation?.isAnimating != true) {
             startCookiesAnimation(context, omnibarViews)
         } else {
             enqueueCookiesAnimation = true
@@ -170,19 +156,6 @@ class BrowserLottieTrackersAnimatorHelper @Inject constructor(
         stopTrackersAnimation()
         stopCookiesAnimation()
         omnibarViews.forEach { it.alpha = 1f }
-    }
-
-    override fun finishPartialTrackerAnimation() {
-        val trackersAnimation = this.trackersAnimation ?: return
-        val shieldAnimation = this.shieldAnimation ?: return
-
-        runPartialAnimation = false
-        completePartialAnimation = true
-
-        trackersAnimation.setMinAndMaxProgress(0.5f, 1f)
-        shieldAnimation.setMinAndMaxProgress(0.5f, 1f)
-        trackersAnimation.playAnimation()
-        shieldAnimation.playAnimation()
     }
 
     private fun tryToStartCookiesAnimation(
