+++ app/src/main/java/com/duckduckgo/app/cta/ui/CtaViewModel.kt
@@ -31,7 +31,6 @@ import com.duckduckgo.app.global.model.domain
 import com.duckduckgo.app.global.model.orderedTrackerBlockedEntities
 import com.duckduckgo.app.onboarding.store.*
 import com.duckduckgo.app.onboarding.ui.page.extendedonboarding.ExtendedOnboardingExperimentVariantManager
-import com.duckduckgo.app.pixels.AppPixelName
 import com.duckduckgo.app.privacy.db.UserAllowListRepository
 import com.duckduckgo.app.settings.db.SettingsDataStore
 import com.duckduckgo.app.statistics.pixels.Pixel
@@ -107,12 +106,6 @@ class CtaViewModel @Inject constructor(
         val wasCleared = activeSurvey != null && survey == null
         activeSurvey = survey
         return wasCleared
-    }
-
-    suspend fun hideTipsForever(cta: Cta) {
-        settingsDataStore.hideTips = true
-        pixel.fire(AppPixelName.ONBOARDING_DAX_ALL_CTA_HIDDEN, cta.pixelCancelParameters())
-        userStageStore.stageCompleted(AppStage.DAX_ONBOARDING)
     }
 
     fun onCtaShown(cta: Cta) {
@@ -201,40 +194,33 @@ class CtaViewModel @Inject constructor(
         }
     }
 
-    suspend fun getExperimentFireDialogCta(): ExperimentOnboardingDaxDialogCta.DaxFireButtonCta? {
+    suspend fun getExperimentFireDialogCta(): OnboardingDaxDialogCta.DaxFireButtonCta? {
         if (!daxOnboardingActive() || daxDialogFireEducationShown()) return null
         return withContext(dispatchers.io()) {
-            return@withContext ExperimentOnboardingDaxDialogCta.DaxFireButtonCta(onboardingStore, appInstallStore)
+            return@withContext OnboardingDaxDialogCta.DaxFireButtonCta(onboardingStore, appInstallStore)
         }
     }
 
-    suspend fun getExperimentSiteSuggestionsDialogCta(): ExperimentOnboardingDaxDialogCta.DaxSiteSuggestionsCta? {
+    suspend fun getExperimentSiteSuggestionsDialogCta(): OnboardingDaxDialogCta.DaxSiteSuggestionsCta? {
         if (!daxOnboardingActive() || !canShowDaxIntroVisitSiteCta()) return null
         return withContext(dispatchers.io()) {
-            return@withContext ExperimentOnboardingDaxDialogCta.DaxSiteSuggestionsCta(onboardingStore, appInstallStore)
+            return@withContext OnboardingDaxDialogCta.DaxSiteSuggestionsCta(onboardingStore, appInstallStore)
         }
     }
 
     private suspend fun getHomeCta(): Cta? {
+        val onboardingEnabled = extendedOnboardingExperimentVariantManager.isAestheticUpdatesEnabled()
         return when {
-            canShowDaxIntroCta() -> {
-                if (extendedOnboardingExperimentVariantManager.isAestheticUpdatesEnabled()) {
-                    ExperimentDaxBubbleOptionsCta.ExperimentDaxIntroSearchOptionsCta(onboardingStore, appInstallStore)
-                } else {
-                    DaxBubbleCta.DaxIntroCta(onboardingStore, appInstallStore)
-                }
+            canShowDaxIntroCta() && onboardingEnabled -> {
+                ExperimentDaxBubbleOptionsCta.ExperimentDaxIntroSearchOptionsCta(onboardingStore, appInstallStore)
             }
 
-            canShowDaxIntroVisitSiteCta() && extendedOnboardingExperimentVariantManager.isAestheticUpdatesEnabled() -> {
+            canShowDaxIntroVisitSiteCta() && onboardingEnabled -> {
                 ExperimentDaxBubbleOptionsCta.ExperimentDaxIntroVisitSiteOptionsCta(onboardingStore, appInstallStore)
             }
 
-            canShowDaxCtaEndOfJourney() -> {
-                if (extendedOnboardingExperimentVariantManager.isAestheticUpdatesEnabled()) {
-                    ExperimentDaxBubbleOptionsCta.ExperimentDaxEndCta(onboardingStore, appInstallStore)
-                } else {
-                    DaxBubbleCta.DaxEndCta(onboardingStore, appInstallStore)
-                }
+            canShowDaxCtaEndOfJourney() && onboardingEnabled -> {
+                ExperimentDaxBubbleOptionsCta.ExperimentDaxEndCta(onboardingStore, appInstallStore)
             }
 
             canShowWidgetCta() -> {
@@ -297,46 +283,33 @@ class CtaViewModel @Inject constructor(
 
             if (!canShowDaxDialogCta()) return null
 
-            // Trackers blocked
-            if (!daxDialogTrackersFoundShown() && !isSerpUrl(it.url) && it.orderedTrackerBlockedEntities().isNotEmpty()) {
-                return if (extendedOnboardingExperimentVariantManager.isAestheticUpdatesEnabled()) {
-                    ExperimentOnboardingDaxDialogCta.DaxTrackersBlockedCta(
+            if (extendedOnboardingExperimentVariantManager.isAestheticUpdatesEnabled()) {
+                // Trackers blocked
+                if (!daxDialogTrackersFoundShown() && !isSerpUrl(it.url) && it.orderedTrackerBlockedEntities().isNotEmpty()) {
+                    return OnboardingDaxDialogCta.DaxTrackersBlockedCta(
                         onboardingStore,
                         appInstallStore,
                         it.orderedTrackerBlockedEntities(),
                     )
-                } else {
-                    DaxDialogCta.DaxTrackersBlockedCta(onboardingStore, appInstallStore, it.orderedTrackerBlockedEntities(), host)
                 }
-            }
 
-            // Is major network
-            if (it.entity != null) {
-                it.entity?.let { entity ->
-                    if (!daxDialogNetworkShown() && DaxDialogCta.mainTrackerNetworks.contains(entity.displayName)) {
-                        return if (extendedOnboardingExperimentVariantManager.isAestheticUpdatesEnabled()) {
-                            ExperimentOnboardingDaxDialogCta.DaxMainNetworkCta(onboardingStore, appInstallStore, entity.displayName, host)
-                        } else {
-                            DaxDialogCta.DaxMainNetworkCta(onboardingStore, appInstallStore, entity.displayName, host)
+                // Is major network
+                if (it.entity != null) {
+                    it.entity?.let { entity ->
+                        if (!daxDialogNetworkShown() && OnboardingDaxDialogCta.mainTrackerNetworks.contains(entity.displayName)) {
+                            return OnboardingDaxDialogCta.DaxMainNetworkCta(onboardingStore, appInstallStore, entity.displayName, host)
                         }
                     }
                 }
-            }
 
-            // SERP
-            if (isSerpUrl(it.url) && !daxDialogSerpShown()) {
-                return if (extendedOnboardingExperimentVariantManager.isAestheticUpdatesEnabled()) {
-                    ExperimentOnboardingDaxDialogCta.DaxSerpCta(onboardingStore, appInstallStore)
-                } else {
-                    DaxDialogCta.DaxSerpCta(onboardingStore, appInstallStore)
+                // SERP
+                if (isSerpUrl(it.url) && !daxDialogSerpShown()) {
+                    return OnboardingDaxDialogCta.DaxSerpCta(onboardingStore, appInstallStore)
                 }
-            }
 
-            if (!isSerpUrl(it.url) && !daxDialogOtherShown() && !daxDialogTrackersFoundShown() && !daxDialogNetworkShown()) {
-                return if (extendedOnboardingExperimentVariantManager.isAestheticUpdatesEnabled()) {
-                    ExperimentOnboardingDaxDialogCta.DaxNoTrackersCta(onboardingStore, appInstallStore)
-                } else {
-                    DaxDialogCta.DaxNoSerpCta(onboardingStore, appInstallStore)
+                // No trackers blocked
+                if (!isSerpUrl(it.url) && !daxDialogOtherShown() && !daxDialogTrackersFoundShown() && !daxDialogNetworkShown()) {
+                    return OnboardingDaxDialogCta.DaxNoTrackersCta(onboardingStore, appInstallStore)
                 }
             }
             return null
@@ -344,8 +317,6 @@ class CtaViewModel @Inject constructor(
     }
 
     private fun daxDialogIntroShown(): Boolean = dismissedCtaDao.exists(CtaId.DAX_INTRO)
-
-    private fun daxDialogIntroVisitSiteShown(): Boolean = dismissedCtaDao.exists(CtaId.DAX_INTRO_VISIT_SITE)
 
     private fun daxDialogEndShown(): Boolean = dismissedCtaDao.exists(CtaId.DAX_END)
 
@@ -361,7 +332,7 @@ class CtaViewModel @Inject constructor(
 
     private fun pulseFireButtonShown(): Boolean = dismissedCtaDao.exists(CtaId.DAX_FIRE_BUTTON_PULSE)
 
-    private fun isSerpUrl(url: String): Boolean = url.contains(DaxDialogCta.SERP)
+    private fun isSerpUrl(url: String): Boolean = url.contains(OnboardingDaxDialogCta.SERP)
 
     private suspend fun daxOnboardingActive(): Boolean = userStageStore.daxOnboardingActive()
 
@@ -411,6 +382,7 @@ class CtaViewModel @Inject constructor(
         }
     }
 
+    @Deprecated("New users won't have this option available since extended onboarding")
     private fun hideTips() = settingsDataStore.hideTips
 
     companion object {
