+++ app/src/main/java/com/duckduckgo/app/settings/db/SettingsDataStore.kt
@@ -34,6 +34,8 @@ import javax.inject.Inject
 interface SettingsDataStore {
 
     var lastExecutedJobId: String?
+
+    @Deprecated(message = "hideTips variable is deprecated and no longer available in onboarding")
     var hideTips: Boolean
     var autoCompleteSuggestionsEnabled: Boolean
     var appIcon: AppIcon
