+++ app/src/main/res/layout/include_new_browser_tab.xml
@@ -29,6 +29,7 @@
         android:id="@+id/browserBackground"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
+        android:background="@drawable/onboarding_experiment_background_bitmap"
         android:paddingTop="?attr/actionBarSize"
         android:foreground="@android:color/transparent">
